Break propka.calculations into parts.
Broken into three files to simplify import analysis:

* `calculations.py` - distance calculations and simple math
* `energy.py` - energy calculations
* `hydrogens.py` - hydrogen geometry calculations

Working towards addressing https://github.com/jensengroup/propka-3.1/issues/49
